Remove status of closed file descriptors
Calling close() on a file descriptor will remove any kevents that
reference the descriptor, but the saved state of that file descriptor
remains. This causes trouble, because fd numbers are reused, and later
you are not able to add an event that references a file descriptor with
the same numerical value as a closed one.