Safely interpose fcntl() on architectures with bounded variadic arguments
`fcntl()` abuses C variadic arguments to implement optional parameters.
We want to forward this argument to the real `fcntl()` call, but doing
so is non-portable. On most architectures, we can just read an
`intptr_t`/`void*` and this will give us a (semi-)valid result even if
no argument/an int was passed instead since `va_arg()` will generally
read the next argument register or arbitrary data from the stack.

On CHERI-enabled architectures, variadic arguments are tightly
bounded, which means that reading a `void*` if an `int` was passed will
result in a runtime trap. To avoid this problem this commit uses a new
`vfcntl` libc function (https://github.com/CTSRD-CHERI/cheribsd/pull/2194),
that takes a va_list instead of variadic arguments.

Since this needs a new HAVE_VCNTL compile-time check, this adds a new
config.h header that also includes the HAVE_TIMERFD value. This removes
the need to pass -D flags on the command line.

This reduces the number of test failures on CheriBSD Morello from 210
to 8.