Check errno on startup in timerfd__{reset_to_very_long,unmodified_errno}
Due to a bug in CheriBSD, errno was set non-zero when starting this test,
but without these checks the test confusingly failed in the later check
which made me think something was wrong in timerfd_create. Check errno on
startup to make this more obvious.

See https://github.com/CTSRD-CHERI/cheribsd/issues/1424